Make it controllable how data is extracted from a Grid
